Refactor AbstractSignatureMatcher to use underscore prefixes for methods
